Some refactoring of the code used for caching. Also added functions to get user role attributes.
